chore: clear App.vue into KirstensView
